Partition tests/conftest.py into fixtures package (#583)
Lift refresh-system + emitter fixtures from the 710-line tests/conftest.py
into tests/fixtures/<concern>.py modules. Conftest becomes ~80 lines:
sys.path inserts, the cross-cutting pact_context fixture (26 consumers),
and re-exports of the pytest-injected fixtures from fixtures.refresh_system.

- New: tests/fixtures/{__init__,emitter,refresh_system}.py
- Conftest 710 → 88 lines; mock_env + cli_memory_dict deleted (zero consumers)
- 10 test files updated: from conftest -> from fixtures.<concern>
- Lazy import of agent_handoff_emitter preserved inside _run_main
- 6985 passed, 8 skipped — exact match pre/post-change

Closes #583
