+++ pact-plugin/tests/test_emitter_happy_and_gates.py
@@ -9,7 +9,7 @@ negative gates.
 """
 import pytest
 
-from conftest import VALID_HANDOFF, _run_main
+from fixtures.emitter import VALID_HANDOFF, _run_main
 
 
 class TestHappyPath:
