+++ pact-plugin/tests/test_emitter_path_sanitization.py
@@ -3,7 +3,7 @@ from pathlib import Path
 
 import pytest
 
-from conftest import VALID_HANDOFF, _run_main
+from fixtures.emitter import VALID_HANDOFF, _run_main
 
 
 class TestPathSanitization:
