+++ pact-plugin/tests/test_emitter_race_regression.py
@@ -10,7 +10,7 @@ phantom-fire-revert sequence.
 """
 import pytest
 
-from conftest import VALID_HANDOFF, _run_main
+from fixtures.emitter import VALID_HANDOFF, _run_main
 
 
 class TestRaceShapeRegression:
